UnifiedStore: moved into store package for consistency (it is used by session, but it is not part of the session, especially when used by state)
