+++ mlte/backend/main.py
@@ -11,8 +11,8 @@ import mlte._private.hosts as util
 import mlte.backend.core.app_factory as app_factory
 from mlte.backend.core.config import settings
 from mlte.backend.core.state import state
-from mlte.session.unified_store import setup_stores
 from mlte.store.base import StoreType, StoreURI
+from mlte.store.unified_store import setup_stores
 
 # Application exit codes
 EXIT_SUCCESS = 0
